Update ikala item to gradle structure
