--- Ikala-Item/.classpath
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" output="bin/classes" path="src"/>
-	<classpathentry kind="src" output="bin/test-classes" path="test">
-		<attributes>
-			<attribute name="test" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.junit.JUNIT_CONTAINER/5"/>
-	<classpathentry kind="lib" path="lib/plugins/Ikala-Core-0.4.0.jar" sourcepath="lib/info/Ikala-Core-0.4.0-sources.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:platform:/resource/Ikala-Item/lib/info/Ikala-Core-0.4.0-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="lib" path="lib/plugins/Ikala-Database-0.0.1.jar" sourcepath="lib/info/Ikala-Database-0.0.1-sources.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:platform:/resource/Ikala-Item/lib/info/Ikala-Database-0.0.1-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="lib" path="lib/plugins/EntityComponentSystem-0.0.1.jar" sourcepath="lib/info/EntityComponentSystem-0.0.1-sources.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:platform:/resource/Ikala-Item/lib/info/EntityComponentSystem-0.0.1-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="output" path="bin/classes"/>
-</classpath>
