--- Ikala-Item/plugin.yml
@@ -1,8 +1,0 @@
-name: Ikala-Item
-version: 0.0.1
-description: An extensible item system
-authors: [maceris]
-main-class: com.ikalagaming.item.ItemPlugin
-dependencies:
-  - Ikala-Database
-  - EntityComponentSystem
