+++ Ikala-Item/settings.gradle
@@ -1,0 +1,5 @@
+rootProject.name = 'Ikala-Item'
+include ':Ikala-Core'
+project(':Ikala-Core').projectDir = new File('../../Ikala-Core')
+include ':EntityComponentSystem'
+project(':EntityComponentSystem').projectDir = new File('../EntityComponentSystem')
