+++ Ikala-Item/src/main/resources/plugin.yml
@@ -1,0 +1,8 @@
+name: Ikala-Item
+version: 0.0.1
+description: An extensible item system
+authors: [maceris]
+main-class: com.ikalagaming.item.ItemPlugin
+dependencies:
+  - Ikala-Database
+  - EntityComponentSystem
